\name{Example}
\alias{Example}
\title{Data Example}

\description{A data example with 1000 observations. The data contains four variables: an continuous outcome variable (y), a binary treatment variable (d), a continuous mediator variable (m) and two continuous covariates (x).
}

\usage{
data(example)
}
